fix(install-targets): dedupe Continue/Windsurf GateGuard wiring, fix import prefixes
SonarCloud flagged continue-home.js/continue-project.js and
windsurf-home.js/windsurf-project.js as duplicated (their gate-wiring
functions were byte-identical). Extracted the shared logic into
continue-gateguard-hooks.js and windsurf-gateguard-operations.js, and
switched the remaining unprefixed fs/os/path requires to their node:
equivalents (codex-home.js, claude-settings-hooks.js,
windsurf-gateguard-hooks.js, windsurf-gateguard-adapter.js), which also
resolved the exception-handling and optional-chaining smells flagged in
windsurf-gateguard-adapter.js.
