+++ scripts/lib/install-targets/continue-project.js
@@ -4,28 +4,8 @@ const {
   createRemappedOperation,
 } = require('./helpers');
 const {
-  createGateGuardScriptCopyOperations,
-  createPreToolUseGateGuardHookMergeOperation,
-} = require('../claude-settings-hooks');
-
-// See continue-home.js for why the Claude Code merge operation is reusable
-// as-is: Continue CLI's PreToolUse hook schema and tool_name values are
-// intentionally Claude Code-compatible, and .continue/settings.json (project)
-// is read at the same precedence tier as .claude/settings.json.
-function createContinueGateGuardOperations(adapter, targetRoot) {
-  const copyOperations = createGateGuardScriptCopyOperations(
-    (moduleId, sourceRelativePath, destinationPath, options) => (
-      createRemappedOperation(adapter, moduleId, sourceRelativePath, destinationPath, options)
-    ),
-    targetRoot
-  );
-
-  const mergeOperations = ['Edit', 'Write', 'MultiEdit', 'Bash'].map(matcher => (
-    createPreToolUseGateGuardHookMergeOperation(targetRoot, matcher)
-  ));
-
-  return [...copyOperations, ...mergeOperations];
-}
+  createContinueGateGuardOperations,
+} = require('../continue-gateguard-hooks');
 
 module.exports = createInstallTargetAdapter({
   id: 'continue-project',
@@ -44,7 +24,7 @@ module.exports = createInstallTargetAdapter({
 
     return [
       ...createFlatSkillPlanOperations(input, adapter),
-      ...createContinueGateGuardOperations(adapter, targetRoot),
+      ...createContinueGateGuardOperations(adapter, targetRoot, createRemappedOperation),
     ];
   },
 });
